+++ src/scenes/ProgressReports/EditForm/ProgressReportDrawer.graphql
