+++ src/scenes/ProgressReports/EditForm/ProgressReportDrawer.tsx
@@ -3,7 +3,7 @@ import { Box, Drawer } from '@mui/material';
 import { useEffect } from 'react';
 import { useMatch } from 'react-router-dom';
 import { useNavigate } from '~/components/Routing';
-import { useProgressReportContext } from '../../ProgressReportContext';
+import { useProgressReportContext } from '../ProgressReportContext';
 import { ProgressReportEditDocument } from './ProgressReportDrawer.graphql';
 import { ProgressReportDrawerHeader } from './ProgressReportDrawerHeader';
 import { ProgressReportSidebar } from './ProgressReportSidebar';
