+++ src/scenes/ProgressReports/EditForm/Steps/CommunityStory/CommunityStoryStep.graphql
