+++ src/scenes/ProgressReports/EditForm/Steps/PromptVariant/VariantResponsesAccordion.tsx
