+++ src/scenes/ProgressReports/EditForm/Steps/TeamHighlight/TeamHighlightStep.graphql
